# APISpec.yml
...
openapi: 3.0.2
info:
  title: ECE 461 - Spring 2023 - Project 2
  version: 2.3.0
  description: "API for ECE 461/Spring 2023/Project 2: A Trustworthy Module Registry"
  termsOfService: http://swagger.io/terms/
  contact:
    name: Prof. Davis
    url: http://davisjam.github.io
    email: davisjam@purdue.edu
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /package/{id}/rate:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageRating'
          description: Return the rating. Only use this if each metric was computed
            successfully.
        "400":
          description: "There is missing field(s) in the PackageID/AuthenticationToken\
            \ or it is formed improperly, or the AuthenticationToken is invalid."
        "404":
          description: Package does not exist.
        "500":
          description: The package rating system choked on at least one of the metrics.
      operationId: PackageRate
    parameters:
    - name: id
      schema:
        $ref: '#/components/schemas/PackageID'
      in: path
      required: true
    - name: X-Authorization
      description: ""
      schema:
        $ref: '#/components/schemas/AuthenticationToken'
      in: header
      required: true
...
components:
  schemas:
...
    PackageID:
      description: ""
      type: string
    PackageRating:
      description: |-
        Package rating (cf. Project 1).

        If the Project 1 that you inherited does not support one or more of the original properties, denote this with the value "-1".
      required:
      - RampUp
      - Correctness
      - BusFactor
      - ResponsiveMaintainer
      - LicenseScore
      - GoodPinningPractice
      - PullRequest
      - NetScore
      type: object
      properties:
        BusFactor:
          format: double
          description: ""
          type: number
        Correctness:
          format: double
          description: ""
          type: number
        RampUp:
          format: double
          description: ""
          type: number
        ResponsiveMaintainer:
          format: double
          description: ""
          type: number
        LicenseScore:
          format: double
          description: ""
          type: number
        GoodPinningPractice:
          format: double
          description: "The fraction of its dependencies that are pinned to at least\
            \ a specific major+minor version, e.g. version 2.3.X of a package. (If\
            \ there are zero dependencies, they should receive a 1.0 rating. If there\
            \ are two dependencies, one pinned to this degree, then they should receive\
            \ a Â½ = 0.5 rating)."
          type: number
        PullRequest:
          format: double
          description: The fraction of project code that was introduced through pull
            requests with a code review.
          type: number
        NetScore:
          format: double
          description: From Part 1
          type: number
...
    AuthenticationToken:
      description: "The spec permits you to use any token format you like. You could,\
        \ for example, look into JSON Web Tokens (\"JWT\", pronounced \"jots\"): https://jwt.io."
      type: string
...
  securitySchemes:
    UserAndPasswordYieldsToken:
      type: apiKey
      description: Use the /authentication endpoint to obtain an authorization token.
      name: X-Authorization
      in: header