# src/model/packageRating.ts
/**
 * ECE 461 - Spring 2023 - Project 2
 * API for ECE 461/Spring 2023/Project 2: A Trustworthy Module Registry
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: davisjam@purdue.edu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import Joi from "joi";
import mongoose from "mongoose";


/**
 * Package rating (cf. Project 1).  If the Project 1 that you inherited does not support one or more of the original properties, denote this with the value \"-1\".
 */
export interface PackageRating {
  NetScore: number;
  /**
   *
   */
  BusFactor: number;
  /**
   *
   */
  Correctness: number;
  /**
   *
   */
  RampUp: number;
  /**
   *
   */
  ResponsiveMaintainer: number;
  /**
   *
   */
  LicenseScore: number;
  /**
   * The fraction of its dependencies that are pinned to at least a specific major+minor version, e.g. version 2.3.X of a package. (If there are zero dependencies, they should receive a 1.0 rating. If there are two dependencies, one pinned to this degree, then they should receive a Â½ = 0.5 rating).
   */
  GoodPinningPractice: number;

  GoodEngineeringPractice: number;
}

export const PackageRatingSchema: mongoose.Schema<PackageRating> = new mongoose.Schema<PackageRating>({
  NetScore: { type: Number, required: true },
  BusFactor: { type: Number, required: true },
  Correctness: { type: Number, required: true },
  RampUp: { type: Number, required: true },
  ResponsiveMaintainer: { type: Number, required: true },
  LicenseScore: { type: Number, required: true },
  GoodPinningPractice: { type: Number, required: true },
  GoodEngineeringPractice: { type: Number, required: true },
});

export const PackageRatingModel = mongoose.model<PackageRating>("PackageRating", PackageRatingSchema);

export const PackageRatingUploadValidation = Joi.object({
  NetScore: Joi.number().min(0.5).required(),
  BusFactor: Joi.number().min(0.5).required(),
  Correctness: Joi.number().min(0.5).required(),
  RampUp: Joi.number().min(0.5).required(),
  ResponsiveMaintainer: Joi.number().min(0.5).required(),
  LicenseScore: Joi.number().min(0.5).required(),
  GoodPinningPractice: Joi.number().min(0.5).required(),
  GoodEngineeringPractice: Joi.number().min(0.5).required(),
});